requirements.txt
# Core framework
crewai==0.80.0
crewai-tools==0.17.0

# LLM
anthropic>=0.40.0

# Data parsing
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0
openpyxl>=3.1.2
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0

# Financial modeling
numpy-financial>=1.0.0

# Output generation
python-pptx>=1.0.0
reportlab>=4.2.0
weasyprint>=62.0
Pillow>=10.2.0
plotly>=5.22.0
kaleido>=0.2.1

# Vector memory / RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Web Dashboard
nicegui>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
tenacity>=8.2.0
loguru>=0.7.2
rich>=13.7.0
typer>=0.12.0
